# source/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/crawler/update/{id}": {
            "put": {
                "description": "Update request form by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Crawler API Request"
                ],
                "summary": "Update request form by ID",
                "operationId": "update-user-request-form-by-id-crawler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated User Request Form Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Form"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/input": {
            "post": {
                "description": "Create new user request form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API Request"
                ],
                "summary": "Create new user request form",
                "operationId": "create-new-user-request-form",
                "parameters": [
                    {
                        "description": "User Request Form Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Form"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/update/{id}": {
            "put": {
                "description": "Update user request form by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API Request"
                ],
                "summary": "Update user request form by ID",
                "operationId": "update-user-request-form-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated User Request Form Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Form"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/{id}": {
            "delete": {
                "description": "Delete user request form by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User API Request"
                ],
                "summary": "Delete user request form by ID",
                "operationId": "delete-user-request-form-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/view_all": {
            "get": {
                "description": "Get all user request forms",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Request"
                ],
                "summary": "Get all user request forms",
                "operationId": "get-all-user-request-forms",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Form"
                            }
                        }
                    }
                }
            }
        },
        "/api/view_detail/{id}": {
            "get": {
                "description": "Get user request form by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API Request"
                ],
                "summary": "Get user request form by ID",
                "operationId": "get-user-request-form-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Form"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Form": {
            "type": "object",
            "properties": {
                "deadline": {
                    "type": "string"
                },
                "division": {
                    "type": "string"
                },
                "inputList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.InputList"
                    }
                },
                "link": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "request_domain": {
                    "type": "string"
                },
                "requester_name": {
                    "type": "string"
                },
                "tag": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.InputList": {
            "type": "object",
            "properties": {
                "input": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "output": {
                    "type": "string"
                }
            }
        }
    }
}